+++ cargo-concordium/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Unreleased changes
 
-- Fix a bug so that a non-zero status code is now returned by the integration tests if they fail to run successfully.
+- Fix a bug so that a non-zero status code is now returned by `cargo concordium test` if tests fail.
 
 ## 3.1.4